moved optionDropdown into its own file converted the mod options button into a statefulwidget to handle changing the color when selection is complete.
